add PostResponseDTO to separate file
