+++ forum/src/main/java/com/grupp4/forum/post/PostController.java
@@ -1,8 +1,7 @@
 package com.grupp4.forum.post;
 
 import com.grupp4.forum.dto.CreatePostDTO;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
+import com.grupp4.forum.dto.PostResponseDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -10,8 +9,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-
-import java.util.UUID;
 
 @RestController
 @RequestMapping("/post")
@@ -30,29 +27,4 @@ public class PostController {
             return ResponseEntity.badRequest().body("Hoppsan här blev det fel (ska va errorResponseDTO");
         }
     }
-
-    @AllArgsConstructor
-    @Getter
-    public static class PostResponseDTO {
-        private UUID id;
-        private String content;
-        private UUID userId;
-        private String username;
-
-        //uncomment när commentcontroller är på plats
-//        private List<CommentController.CommentResponseDTO> comments;
-
-        public static PostResponseDTO fromModel(Post post) {
-            return new PostResponseDTO(
-                    post.getId(),
-                    post.getContent(),
-                    post.getUser().getId(),
-                    post.getUser().getName()
-
-                    //uncomment när commentcontroller är på plats
-//                  post.getComments().stream().map(CommentController.CommentResponseDTO::fromModel).toList()
-            );
-        }
-    }
-
 }
